Small refactoring of class names and namespaces

+++ Platform/Arbiter/Contracts/IGameManager.cs
@@ -7,6 +7,6 @@ namespace Arbiter.ServiceContracts
     public interface IGameManager
     {
         [OperationContract]
-        void TeamReady(int TemaId);
+        void TeamReady(byte TemaId);
     }
 }
